Added MorphColors and fixed JSONLoader

+++ lib/src/loaders/Loader.dart
@@ -207,7 +207,10 @@ class Loader {
 
   }
 
-  _rgb2hex( rgb ) => ( rgb[ 0 ] * 255 << 16 ) + ( rgb[ 1 ] * 255 << 8 ) + rgb[ 2 ] * 255;
+  _rgb2hex( rgb ) =>
+      ( (rgb[ 0 ] * 255).toInt() << 16 ) +
+      ( (rgb[ 1 ] * 255).toInt() << 8 ) +
+      rgb[ 2 ].toInt() * 255;
 
 
   _createMaterial( Map m, String texturePath ) {
